+++ examples/MemoryBenchmark/generate_readme.py
@@ -38,7 +38,7 @@ by the runtime environment of the processor. For example, it often seems like
 the ESP8266 allocates flash memory in blocks of a certain quantity, so the
 calculated flash size can jump around in unexpected ways.
 
-**Version**: AceTMI v0.4
+**Version**: AceTMI v0.5
 
 **DO NOT EDIT**: This file was auto-generated using `make README.md`.
 
@@ -90,8 +90,8 @@ $ make README.md
 The following shows the flash and static memory sizes of the `MemoryBenchmark`
 program for various `LedModule` configurations and various Writer classes.
 
-* `SimpleTmiInterface`: implementation using `digitalWrite()`
-* `SimpleTmiFastInterface`: implementation using `digitalWriteFast()`
+* `SimpleTmi1637Interface`: implementation using `digitalWrite()`
+* `SimpleTmi1637FastInterface`: implementation using `digitalWriteFast()`
 
 ### ATtiny85