+++ src/AceTMI.h
@@ -37,8 +37,8 @@ SOFTWARE.
 #define ACE_TMI_H
 
 // Version format: xxyyzz == "xx.yy.zz"
-#define ACE_TMI_VERSION 400
-#define ACE_TMI_VERSION_STRING "0.4"
+#define ACE_TMI_VERSION 500
+#define ACE_TMI_VERSION_STRING "0.5"
 
 // Blacklist platforms using https://github.com/arduino/ArduinoCore-api due to
 // incompatibilities.
@@ -47,12 +47,12 @@ SOFTWARE.
 #endif
 
 // Files exported by this main header file.
-#include "ace_tmi/SimpleTmiInterface.h"
+#include "ace_tmi/SimpleTmi1637Interface.h"
 #include "ace_tmi/SimpleTmi1638Interface.h"
 
 // The following are commented out because they work only on AVR platforms with
 // a suitable <digitalWriteFast.h> library.
-//#include "ace_tmi/SimpleTmiFastInterface.h"
+//#include "ace_tmi/SimpleTmi1637FastInterface.h"
 //#include "ace_tmi/SimpleTmi1638FastInterface.h"
 
 #endif